+++ bindings/nostr-sdk-ffi/src/pool/mod.rs
@@ -67,7 +67,7 @@ impl RelayPool {
         self.inner.blacklist().into()
     }
 
-    /// Get relays
+    /// Get relays with `READ` or `WRITE` flags
     pub async fn relays(&self) -> HashMap<String, Arc<Relay>> {
         self.inner
             .relays()